Tidy everything up for submission
